backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
langchain>=0.1.0
langchain-community>=0.1.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
langgraph>=0.1
openai>=0.27
faster-whisper>=0.6
transformers>=4.40
torch>=2.0
chromadb>=0.4
qdrant-client>=1.7
pydantic>=1.10
httpx>=0.24
python-multipart>=0.0.6
pydub>=0.25
python-dotenv>=1.0
pytest>=7.0
structlog>=23.0
beautifulsoup4==4.13.5
fastapi==0.116.1
html5lib==1.1
jinja2==3.1.6
langchain==0.3.27
langchain-astradb==0.6.1
langchain_core==0.3.75
langchain-google-genai==2.1.8
langchain-groq==0.3.6
lxml==6.0.1
python-dotenv==1.1.1
python-multipart==0.0.20
selenium==4.35.0
streamlit==1.49.1
undetected-chromedriver==3.5.5
uvicorn==0.35.0
structlog==25.4.0
langgraph==0.6.7
